cmake/library.cmake: Remove pc.${PROJECT_NAME} only if it exists

+++ cmake/library.cmake
@@ -798,8 +798,14 @@ endif()
       #
       # Create a pc.${PROJECT_NAME} directory
       #
+      if(EXISTS ${CMAKE_CURRENT_BINARY_DIR}/pc.${PROJECT_NAME})
+        execute_process(
+          COMMAND ${CMAKE_COMMAND} -E rm -rf ${CMAKE_CURRENT_BINARY_DIR}/pc.${PROJECT_NAME}
+	  ${_cmake_helpers_process_command_echo_stdout}
+          COMMAND_ERROR_IS_FATAL ANY
+        )
+      endif()
       execute_process(
-        COMMAND ${CMAKE_COMMAND} -E rm -rf ${CMAKE_CURRENT_BINARY_DIR}/pc.${PROJECT_NAME}
         COMMAND ${CMAKE_COMMAND} -E make_directory ${CMAKE_CURRENT_BINARY_DIR}/pc.${PROJECT_NAME}
 	${_cmake_helpers_process_command_echo_stdout}
         COMMAND_ERROR_IS_FATAL ANY
@@ -815,8 +821,8 @@ endif()
 	message(STATUS "[${_cmake_helpers_logprefix}] Generating ${_cmake_helpers_library_pc_in}")
       endif()
       file(CONFIGURE
-      OUTPUT ${_cmake_helpers_library_pc_in}
-      CONTENT [[prefix=${pcfiledir}/../..
+        OUTPUT ${_cmake_helpers_library_pc_in}
+        CONTENT [[prefix=${pcfiledir}/../..
 exec_prefix=${prefix}
 bindir=${exec_prefix}/@CMAKE_HELPERS_INSTALL_BINDIR@
 includedir=${prefix}/@CMAKE_HELPERS_INSTALL_INCLUDEDIR@
